Refactor asset API helpers
- Move `resource_file`, `is_within_directory`, and `is_directory_in_archive` to new **`plantdb/server/api/utils.py`**
- Clean up unused imports in `assets.py` (e.g., `os`, `pathlib`, `BytesIO`, `ZipFile`, `numpy`, `requests`)
- Import `resource_file` from utils in `assets.py`
- Delete the original helper implementations from `assets.py` to avoid duplication
- Keep existing functionality unchanged while improving module organization.
