+++ src/server/plantdb/server/api/assets.py
@@ -93,17 +93,11 @@ Circle
 import json
 import logging
 import mimetypes
-import os
-import pathlib
 import warnings
 from http.client import HTTPException
-from io import BytesIO
 from pathlib import Path
-from zipfile import ZipFile
 
-import numpy as np
 import pybase64
-import requests
 from flask import after_this_request
 from flask import jsonify
 from flask import make_response
@@ -120,11 +114,11 @@ from plantdb.commons.fsdb.exceptions import ScanNotFoundError
 from plantdb.commons.io import read_json
 from plantdb.commons.log import get_logger
 from plantdb.server import webcache
+from plantdb.server.api.utils import resource_file
 from plantdb.server.core.security import add_jwt_from_header
 from plantdb.server.core.security import rate_limit
 from plantdb.server.core.security import sanitize_ids
 from plantdb.server.core.security import use_guest_as_default
-from plantdb.server.core.utils import compute_fileset_matches
 from plantdb.server.services.assets import ArchiveError
 from plantdb.server.services.assets import ExtractionError
 from plantdb.server.services.assets import FileUploadError
@@ -530,59 +524,6 @@ class Image(Resource):
             resp.headers["X-Content-Encoding"] = "binary"
 
         return resp
-
-
-def resource_file(db, scan_id, task_name, **kwargs):
-    """Retrieve a specific ``File`` object from the database.
-
-    Parameters
-    ----------
-    db : plantdb.commons.fsdb.core.FSDB
-        The database instance.
-    scan_id : str
-        Identifier of the scan containing the requested file.
-    task_name : str
-        Name of the task (fileset) and file to fetch.
-    **kwargs
-        Additional arguments passed to ``FSDB.get_scan`` (e.g. JWT token).
-
-    Returns
-    -------
-    tuple
-        Either ``(file, 200)`` on success or ``(error_dict, status_code)`` on failure.
-        The error dictionary always uses the key ``"error"`` for consistency.
-    """
-
-    # Get the corresponding `Scan` instance
-    try:
-        scan = db.get_scan(scan_id, **kwargs)
-
-    except NoAuthUserError as e:
-        return {'error': str(e)}, 401  # HTTP 401 Unauthorized (authentication)
-    except ScanNotFoundError:
-        return {"error": f"Scan '{scan_id}' not found!"}, 400
-
-    task_fs_map = compute_fileset_matches(scan)
-    # Get the corresponding `Fileset` instance
-    try:
-        fs = scan.get_fileset(task_fs_map[task_name])
-    except KeyError:
-        return {"error": f"No fileset mapped for task '{task_name}'."}, 404
-    except FilesetNotFoundError:
-        return {"error": f"Fileset for task '{task_name}' not found."}, 404
-
-    # Get the `File` corresponding to the resource
-    try:
-        file = fs.get_file(task_name)
-    except FileNotFoundError:
-        return {"error": f"File '{fs.id}/{task_name}' not found."}, 404
-    except Exception as exc:                     # Unexpected internal error
-        # Use JSON‑serializable payload; Flask will handle conversion.
-        return {"error": f"Internal server error: {str(exc)}"}, 500
-
-    # Success – return the File object (Flask‑RESTful resources expect the
-    # object itself; the caller can decide the HTTP status if required).
-    return file
 
 
 class PointCloud(Resource):
@@ -1108,54 +1049,6 @@ class Sequence(Resource):
             return measures[type]
         else:
             return measures
-
-
-def is_within_directory(directory, target):
-    """Check if a target path is within a directory.
-
-    This function determines if the absolute path of the target is located
-    within the absolute path of the directory. It uses `os.path.commonpath`
-    to perform the comparison.
-
-    Parameters
-    ----------
-    directory : str or pathlib.Path
-        The path to the directory to check against.
-    target : str or pathlib.Path
-        The path to the target to check if it resides within the directory.
-
-    Returns
-    -------
-    bool
-        ``True`` if the target path is within the directory, ``False`` otherwise.
-    """
-    abs_directory = os.path.abspath(directory)
-    abs_target = os.path.abspath(target)
-    return os.path.commonpath([abs_directory]) == os.path.commonpath([abs_directory, abs_target])
-
-
-def is_directory_in_archive(archive_path, target_dir):
-    """Check if a specific directory exists within an archive file.
-
-    This function checks whether a given directory is present at the top level of a ZIP archive.
-
-    Parameters
-    ----------
-    archive_path : str or pathlib.Path
-        The path to the ZIP archive file.
-    target_dir : str
-        The name of the target directory to check for within the archive.
-
-    Returns
-    -------
-    bool
-        True if the target directory exists at the top level of the archive, False otherwise.
-    """
-    with ZipFile(archive_path, 'r') as zip_ref:
-        # List all members in the zip file
-        top_level_members = [name for name in zip_ref.namelist() if '/' not in name]
-        # Check if the target directory is among them
-        return f"{target_dir}/" in top_level_members or target_dir in top_level_members
 
 
 class Archive(Resource):
